rename eventmesh-storage-standalone packages
