+++ eventmesh-runtime/src/test/java/org/apache/eventmesh/runtime/util/EventMeshUtilTest.java
@@ -22,7 +22,7 @@ import org.apache.eventmesh.common.ThreadPoolFactory;
 import org.apache.eventmesh.common.exception.EventMeshException;
 import org.apache.eventmesh.common.protocol.tcp.EventMeshMessage;
 import org.apache.eventmesh.common.protocol.tcp.UserAgent;
-import org.apache.eventmesh.connector.standalone.broker.model.TopicMetadata;
+import org.apache.eventmesh.storage.standalone.broker.model.TopicMetadata;
 import org.apache.eventmesh.runtime.constants.EventMeshConstants;
 
 import org.apache.http.client.utils.URIBuilder;
