+++ eventmesh-storage/eventmesh-storage-rabbitmq/src/test/java/org/apache/eventmesh/storage/rabbitmq/RabbitmqMockConnectionFactory.java
@@ -15,7 +15,7 @@
  *  limitations under the License.
  */
 
-package org.apache.eventmesh.connector.rabbitmq;
+package org.apache.eventmesh.storage.rabbitmq;
 
 import org.apache.eventmesh.storage.rabbitmq.client.RabbitmqConnectionFactory;
