+++ eventmesh-storage/eventmesh-storage-standalone/src/main/java/org/apache/eventmesh/storage/standalone/broker/StandaloneBroker.java
@@ -15,11 +15,11 @@
  * limitations under the License.
  */
 
-package org.apache.eventmesh.connector.standalone.broker;
+package org.apache.eventmesh.storage.standalone.broker;
 
-import org.apache.eventmesh.connector.standalone.broker.model.MessageEntity;
-import org.apache.eventmesh.connector.standalone.broker.model.TopicMetadata;
-import org.apache.eventmesh.connector.standalone.broker.task.HistoryMessageClearTask;
+import org.apache.eventmesh.storage.standalone.broker.model.MessageEntity;
+import org.apache.eventmesh.storage.standalone.broker.model.TopicMetadata;
+import org.apache.eventmesh.storage.standalone.broker.task.HistoryMessageClearTask;
 
 import org.apache.commons.lang3.tuple.Pair;
