+++ eventmesh-storage/eventmesh-storage-standalone/src/main/java/org/apache/eventmesh/storage/standalone/broker/task/HistoryMessageClearTask.java
@@ -15,12 +15,12 @@
  * limitations under the License.
  */
 
-package org.apache.eventmesh.connector.standalone.broker.task;
+package org.apache.eventmesh.storage.standalone.broker.task;
 
 import org.apache.eventmesh.common.utils.ThreadUtils;
-import org.apache.eventmesh.connector.standalone.broker.MessageQueue;
-import org.apache.eventmesh.connector.standalone.broker.model.MessageEntity;
-import org.apache.eventmesh.connector.standalone.broker.model.TopicMetadata;
+import org.apache.eventmesh.storage.standalone.broker.MessageQueue;
+import org.apache.eventmesh.storage.standalone.broker.model.MessageEntity;
+import org.apache.eventmesh.storage.standalone.broker.model.TopicMetadata;
 
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.TimeUnit;
