+++ eventmesh-storage/eventmesh-storage-standalone/src/main/resources/META-INF/eventmesh/org.apache.eventmesh.api.admin.Admin
@@ -13,4 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-standalone=org.apache.eventmesh.connector.standalone.admin.StandaloneAdminAdaptor
+standalone=org.apache.eventmesh.storage.standalone.admin.StandaloneAdminAdaptor
