+++ eventmesh-storage/eventmesh-storage-standalone/src/main/resources/META-INF/eventmesh/org.apache.eventmesh.api.consumer.Consumer
@@ -13,4 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-standalone=org.apache.eventmesh.connector.standalone.consumer.StandaloneConsumerAdaptor
+standalone=org.apache.eventmesh.storage.standalone.consumer.StandaloneConsumerAdaptor
