+++ eventmesh-storage/eventmesh-storage-standalone/src/main/resources/META-INF/eventmesh/org.apache.eventmesh.api.producer.Producer
@@ -13,4 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-standalone=org.apache.eventmesh.connector.standalone.producer.StandaloneProducerAdaptor
+standalone=org.apache.eventmesh.storage.standalone.producer.StandaloneProducerAdaptor
